Refactor destroy command: remove delete, add safety warnings, enforce --targets
- Remove delete command entirely (code, tests, docs, CLI registration)
- Replace delete with destroy in integration test cleanup steps
- Make --targets required for destroy (no implicit all-stages)
- Add WARNING banner: destroy deletes all matching resources including
  user-created ones, not only CI/CD-created resources
- Add irreversibility warning in CLI docs
- Remove explicit workflow run stopping from destruction phase;
  MWAA Serverless auto-terminates runs on workflow deletion
- Fetch workflow definitions from MWAA API (get_workflow) instead
  of scanning S3 for YAML files
- Add deploy/destroy drift detection: DEPLOY_RESOURCE_TYPES registry
  in resource_types.py with unit test asserting parity with
  DESTROY_SUPPORTED_RESOURCE_TYPES
- Rename --stages to --targets across all commands (including create)
  and all documentation for consistency
- Fix --targets typo in AWS EventBridge put-targets doc examples
- Consolidate cli-commands.md into single 1-10 command list, remove
  duplicate sections, fix numbering, remove stray content
- Split destroy.py (1656 lines) into four modules:
  destroy.py (282), destroy_models.py (68), destroy_validator.py (568),
  destroy_executor.py (382)
- Update specs (requirements, design, tasks) to reflect all changes
