+++ src/smus_cicd/cli.py
@@ -12,7 +12,6 @@ from rich.console import Console
 from . import __version__
 from .commands.bundle import bundle_command
 from .commands.create import create_command_with_output
-from .commands.delete import delete_command
 from .commands.deploy import deploy_command
 
 # Import command functions
@@ -124,7 +123,7 @@ def version():
 # Register commands with proper ordering
 @app.command(
     "describe",
-    help="1. Describe and validate pipeline manifest file. Example: aws-smus-cicd-cli describe --manifest pipeline.yaml --targets dev",
+    help="2. Describe and validate pipeline manifest file. Example: aws-smus-cicd-cli describe --manifest pipeline.yaml --targets dev",
     rich_help_panel="Pipeline Commands",
 )
 def describe(
@@ -161,7 +160,7 @@ def describe(
 
 @app.command(
     "bundle",
-    help="2. Create bundle zip files. Bundles from dev target by default. Example: aws-smus-cicd-cli bundle --targets test",
+    help="3. Create bundle zip files. Bundles from dev target by default. Example: aws-smus-cicd-cli bundle --targets test",
     rich_help_panel="Pipeline Commands",
 )
 def bundle(
@@ -210,7 +209,7 @@ def bundle(
 
 @app.command(
     "deploy",
-    help="3. Deploy bundle files to target (auto-initializes if needed). Example: aws-smus-cicd-cli deploy --manifest pipeline.yaml --targets prod",
+    help="4. Deploy bundle files to target (auto-initializes if needed). Example: aws-smus-cicd-cli deploy --manifest pipeline.yaml --targets prod",
     rich_help_panel="Pipeline Commands",
 )
 def deploy(
@@ -294,7 +293,7 @@ def deploy(
 
 @app.command(
     "monitor",
-    help="4. Monitor workflow status. Example: aws-smus-cicd-cli monitor --manifest pipeline.yaml --targets dev",
+    help="5. Monitor workflow status. Example: aws-smus-cicd-cli monitor --manifest pipeline.yaml --targets dev",
     rich_help_panel="Pipeline Commands",
 )
 def monitor(
@@ -324,7 +323,7 @@ def monitor(
 
 @app.command(
     "create",
-    help="0. Create new pipeline manifest. Example: aws-smus-cicd-cli create --output pipeline.yaml --name 'MyPipeline'",
+    help="1. Create new pipeline manifest. Example: aws-smus-cicd-cli create --output pipeline.yaml --name 'MyPipeline'",
     rich_help_panel="Pipeline Commands",
 )
 def create(
@@ -348,8 +347,8 @@ def create(
     ),
     stages: str = typer.Option(
         "dev,test,prod",
-        "--stages",
-        help="Comma-separated list of stages to create targets for",
+        "--targets",
+        help="Comma-separated list of target stages to create in the manifest",
     ),
     region: str = typer.Option("us-east-1", "--region", help="AWS region"),
 ):
@@ -371,7 +370,7 @@ def create(
 
 @app.command(
     "logs",
-    help="5. Fetch workflow logs from CloudWatch. Example: aws-smus-cicd-cli logs --workflow arn:aws:airflow-serverless:us-east-2:123456789012:workflow/my-workflow",
+    help="7. Fetch workflow logs from CloudWatch. Example: aws-smus-cicd-cli logs --workflow arn:aws:airflow-serverless:us-east-2:123456789012:workflow/my-workflow",
     rich_help_panel="Pipeline Commands",
 )
 def logs(
@@ -425,7 +424,7 @@ def run(
 
 @app.command(
     "test",
-    help="6. Run tests for pipeline targets. Example: aws-smus-cicd-cli test --targets marketing-test-stage",
+    help="8. Run tests for pipeline targets. Example: aws-smus-cicd-cli test --targets marketing-test-stage",
     rich_help_panel="Pipeline Commands",
 )
 def test(
@@ -456,36 +455,8 @@ def test(
 
 
 @app.command(
-    "delete",
-    help="7. Delete projects and environments. Example: aws-smus-cicd-cli delete --targets marketing-test-stage --force",
-    rich_help_panel="Pipeline Commands",
-)
-def delete(
-    pipeline: str = typer.Option(
-        "manifest.yaml", "--manifest", "-m", help="Path to application manifest file"
-    ),
-    targets: str = typer.Option(
-        None,
-        "--targets",
-        "-t",
-        help="Target name(s) - single target or comma-separated list",
-    ),
-    force: bool = typer.Option(False, "--force", "-f", help="Skip confirmation prompt"),
-    async_mode: bool = typer.Option(
-        False, "--async", help="Don't wait for deletion to complete"
-    ),
-    output: str = typer.Option(
-        "TEXT", "--output", "-o", help="Output format: TEXT (default) or JSON"
-    ),
-):
-    """Delete projects and environments that were deployed during initialize."""
-    configure_logging(output, LOG_LEVEL)
-    delete_command(pipeline, targets, force, async_mode, output)
-
-
-@app.command(
     "destroy",
-    help="8. Destroy all resources deployed by the manifest. Example: aws-smus-cicd-cli destroy --manifest manifest.yaml --targets dev --force",
+    help="10. Destroy all resources deployed by the manifest. Example: aws-smus-cicd-cli destroy --manifest manifest.yaml --targets dev --force",
     rich_help_panel="Pipeline Commands",
 )
 def destroy(
@@ -493,10 +464,10 @@ def destroy(
         "manifest.yaml", "--manifest", "-m", help="Path to application manifest file"
     ),
     targets: str = typer.Option(
-        None,
+        ...,
         "--targets",
         "-t",
-        help="Target name(s) - single target or comma-separated list",
+        help="Target name(s) - single target or comma-separated list (required)",
     ),
     force: bool = typer.Option(False, "--force", "-f", help="Skip confirmation prompt"),
     output: str = typer.Option(
@@ -510,7 +481,7 @@ def destroy(
 
 @app.command(
     "integrate",
-    help="Integrate SMUS CI/CD CLI with other tools (Q CLI). Example: aws-smus-cicd-cli integrate qcli",
+    help="9. Integrate SMUS CI/CD CLI with other tools (Q CLI). Example: aws-smus-cicd-cli integrate qcli",
     rich_help_panel="Pipeline Commands",
 )
 def integrate(
