--- tests/integration/delete_test_app/manifest.yaml
@@ -1,60 +1,0 @@
-applicationName: DeleteTestBundle
-content:
-  storage:
-  - connectionName: default.s3_shared
-    append: false
-    include:
-    - src
-    exclude:
-    - .ipynb_checkpoints/
-    - __pycache__/
-    - '*.pyc'
-    name: code
-  - connectionName: default.s3_shared
-    append: true
-    include:
-    - workflows
-    exclude:
-    - .ipynb_checkpoints/
-    - __pycache__/
-    - '*.pyc'
-    name: workflows
-stages:
-  dev:
-    stage: DEV
-    domain:
-      tags:
-        purpose: smus-cicd-testing
-      region: ${DEV_DOMAIN_REGION}
-    project:
-      name: dev-marketing
-  delete-test:
-    stage: TEST
-    domain:
-      tags:
-        purpose: smus-cicd-testing
-      region: ${DEV_DOMAIN_REGION}
-    project:
-      name: delete-test-project
-      create: true
-      profileName: All capabilities
-      role:
-        arn: arn:aws:iam::${AWS_ACCOUNT_ID}:role/Admin
-      owners:
-      - Eng1
-      - arn:aws:iam::${AWS_ACCOUNT_ID}:role/Admin
-      - arn:aws:iam::${AWS_ACCOUNT_ID}:role/GitHubActionsRole-SMUS-CLI-Tests
-      contributors: []
-      userParameters:
-      - EnvironmentConfigurationName: Lakehouse Database
-        parameters:
-        - name: glueDbName
-          value: delete_test_db
-    deployment_configuration:
-      storage:
-      - connectionName: default.s3_shared
-        name: code
-        targetDirectory: src
-      - connectionName: default.s3_shared
-        name: workflows
-        targetDirectory: workflows
